INS-795: add missed pieces for working sync heavy messages over message bus
